Refactoring files structure
- move all views to special dir
