+++ js/views/greeting-view.js
