+++ js/views/modal-confirm-view.js
